requirements.txt
beautifulsoup4==4.12.3
markdownify==0.13.1
ollama==0.3.1
PyPDF2==3.0.1
PyYAML==6.0.1
setuptools==69.5.1
tqdm==4.66.4
